pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.3",
    "imageio-ffmpeg>=0.6.0",
    "python-dotenv>=1.1.1",
    "yt-dlp>=2025.8.27",
]
